Just cosmetics (#170)
* remove unused reference from FederatedPeg.Tests

* Interface CrossChainTransfer
