+++ src/Stratis.FederatedPeg.Features.FederationGateway/TargetChain/CrossChainTransfer.cs
@@ -1,77 +1,49 @@
 ﻿using NBitcoin;
 using Stratis.Bitcoin.Utilities;
+using Stratis.FederatedPeg.Features.FederationGateway.Interfaces;
 
 namespace Stratis.FederatedPeg.Features.FederationGateway.TargetChain
 {
     /// <summary>
-    /// Cross-chain transfer statuses.
-    /// </summary>
-    public enum CrossChainTransferStatus
-    {
-        Partial = 'P',
-        FullySigned = 'F',
-        SeenInBlock = 'S',
-        Rejected = 'R'
-    }
-
-    /// <summary>
     /// Tracks the status of the cross-chain transfer.
     /// </summary>
-    public class CrossChainTransfer : IBitcoinSerializable
+    public class CrossChainTransfer : ICrossChainTransfer
     {
-        /// <summary>
-        /// The transaction id of the deposit transaction.
-        /// </summary>
-        private uint256 depositTransactionId;
+        /// <inheritdoc />
         public uint256 DepositTransactionId => this.depositTransactionId;
+        private uint256 depositTransactionId;
 
-        /// <summary>
-        /// Indicated whether the deposit fields contain information.
-        /// </summary>
-        private bool depositPresent => this.depositTargetAddress != null;
+        /// <inheritdoc />
         public bool DepositPresent => this.depositPresent;
+        private bool depositPresent => this.depositTargetAddress != null;
 
-        /// <summary>
-        /// The block height of the deposit transaction.
-        /// </summary>
+        /// <inheritdoc />
         public long DepositBlockHeight => this.depositBlockHeight;
         private long depositBlockHeight;
 
-        /// <summary>
-        /// The target address of the deposit transaction.
-        /// </summary>
+        /// <inheritdoc />
         public Script DepositTargetAddress => this.depositTargetAddress;
         private Script depositTargetAddress;
 
-        /// <summary>
-        /// The amount (in satoshis) of the deposit transaction.
-        /// </summary>
+        /// <inheritdoc />
         public long DepositAmount => this.depositAmount;
         private long depositAmount;
 
-        /// <summary>
-        /// The unsigned partial transaction containing a full set of available UTXO's.
-        /// </summary>
+        /// <inheritdoc />
         public Transaction PartialTransaction => this.partialTransaction;
         private Transaction partialTransaction;
 
-        /// <summary>
-        /// The hash of the block where the transaction resides on our chain.
-        /// </summary>
+        /// <inheritdoc />
         public uint256 BlockHash => this.blockHash;
         private uint256 blockHash;
 
-        /// <summary>
-        /// The height of the block where the transaction resides on our chain.
-        /// </summary>
+        /// <inheritdoc />
         public int BlockHeight => this.blockHeight;
         private int blockHeight;
 
-        /// <summary>
-        /// The status of the cross chain transfer transaction.
-        /// </summary>
-        private CrossChainTransferStatus status;
+        /// <inheritdoc />
         public CrossChainTransferStatus Status => this.status;
+        private CrossChainTransferStatus status;
 
         /// <summary>
         /// Parameter-less constructor for (de)serialization.
@@ -140,10 +112,7 @@ namespace Stratis.FederatedPeg.Features.FederationGateway.TargetChain
             }
         }
 
-        /// <summary>
-        /// Depending on the status some fields can't be null.
-        /// </summary>
-        /// <returns><c>false</c> if the object is invalid and <c>true</c> otherwise.</returns>
+        /// <inheritdoc />
         public bool IsValid()
         {
             if (this.status == CrossChainTransferStatus.Partial || this.status == CrossChainTransferStatus.SeenInBlock)
@@ -159,10 +128,7 @@ namespace Stratis.FederatedPeg.Features.FederationGateway.TargetChain
             return this.depositTransactionId != null;
         }
 
-        /// <summary>
-        /// Sets the status and verifies that the status is valid given the available fields.
-        /// </summary>
-        /// <param name="status">The new status.</param>
+        /// <inheritdoc />
         public void SetStatus(CrossChainTransferStatus status)
         {
             this.status = status;
@@ -170,11 +136,7 @@ namespace Stratis.FederatedPeg.Features.FederationGateway.TargetChain
             Guard.Assert(IsValid());
         }
 
-        /// <summary>
-        /// Combines signatures from partial transactions received from other federation members.
-        /// </summary>
-        /// <param name="network">The network targeted by the transactions.</param>
-        /// <param name="partials">Partial transactions received from other federation members.</param>
+        /// <inheritdoc />
         public void CombineSignatures(Network network, Transaction[] partials)
         {
             Guard.Assert(this.status == CrossChainTransferStatus.Partial);
